fix accidental promise wrap
The callbacks of promise.{then,catch} are still wrapped in the
promise's try/catch block. We break out of it to prevent unexpected
behavior, like in the case of

```js
module.exports({}, {}, function (err) {
  if (err) throw err
  console.log('ok')
})
```

The `throw err` will be caught by the promise and instead of the
error being properly thrown, we get an uncaught promise warning.